quick update of applications table + bountyassignment api
